+++ src/main/kotlin/Bukkit.kt
@@ -1,13 +1,17 @@
 package ru.endlesscode.bukkitgradle
 
 import org.gradle.api.provider.Provider
-import ru.endlesscode.bukkitgradle.meta.extension.PluginMeta
+import ru.endlesscode.bukkitgradle.plugin.extension.PluginConfiguration
 import ru.endlesscode.bukkitgradle.server.extension.ServerConfiguration
 
 public interface Bukkit {
 
-    /** Plugin meta. */
-    public val meta: PluginMeta
+    /** Plugin plugin. */
+    public val plugin: PluginConfiguration
+
+    @Deprecated("Use 'plugin' field instead", ReplaceWith("plugin"))
+    public val meta: PluginConfiguration
+        get() = plugin
 
     /** Dev server configuration. */
     public val server: ServerConfiguration
@@ -16,5 +20,7 @@ public interface Bukkit {
     public val apiVersion: Provider<String>
 
     /** Plugin Meta generation enabled. */
+    @Deprecated("Use 'plugin.generatePluginYaml' instead", ReplaceWith("plugin.generatePluginYaml"))
     public val generateMeta: Provider<Boolean>
+        get() = plugin.generatePluginYaml
 }