+++ src/main/kotlin/BukkitExtension.kt
@@ -4,20 +4,19 @@ import org.gradle.api.Action
 import org.gradle.api.Project
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
-import org.gradle.api.provider.Provider
 import org.gradle.kotlin.dsl.assign
 import org.gradle.kotlin.dsl.getByType
 import org.gradle.kotlin.dsl.property
 import org.slf4j.LoggerFactory
 import ru.endlesscode.bukkitgradle.extensions.finalizedOnRead
 import ru.endlesscode.bukkitgradle.extensions.warnSyntaxChanged
-import ru.endlesscode.bukkitgradle.meta.extension.PluginMetaImpl
+import ru.endlesscode.bukkitgradle.plugin.extension.PluginConfigurationImpl
 import ru.endlesscode.bukkitgradle.server.ServerConstants
 import ru.endlesscode.bukkitgradle.server.extension.ServerConfigurationImpl
 
 // TODO 1.0: Remove deprecated fields on release
 public open class BukkitExtension internal constructor(
-    public final override val meta: PluginMetaImpl,
+    public final override val plugin: PluginConfigurationImpl,
     public final override val server: ServerConfigurationImpl,
     objects: ObjectFactory,
 ) : Bukkit {
@@ -25,12 +24,6 @@ public open class BukkitExtension internal constructor(
     public final override val apiVersion: Property<String> = objects.property<String>()
         .convention(ServerConstants.DEFAULT_VERSION)
         .finalizedOnRead()
-
-    private val _generateMeta: Property<Boolean> = objects.property<Boolean>()
-        .convention(true)
-        .finalizedOnRead()
-
-    public final override val generateMeta: Provider<Boolean> = _generateMeta
 
     private val logger = LoggerFactory.getLogger("BukkitExtension")
 
@@ -51,8 +44,13 @@ public open class BukkitExtension internal constructor(
         body.execute(server)
     }
 
-    public fun meta(body: Action<PluginMetaImpl>) {
-        body.execute(meta)
+    public fun plugin(body: Action<PluginConfigurationImpl>) {
+        body.execute(plugin)
+    }
+
+    @Deprecated("Use 'plugin' instead", ReplaceWith("plugin(body)"))
+    public fun meta(body: Action<PluginConfigurationImpl>) {
+        plugin(body)
     }
 
     @Deprecated("Use apiVersion instead of version.", ReplaceWith("apiVersion = version"))
@@ -61,9 +59,13 @@ public open class BukkitExtension internal constructor(
         apiVersion = version
     }
 
-    /** Disabled plugin.yml generation. */
+    /** Disables plugin.yml generation. */
+    @Deprecated(
+        "Use 'plugin.disablePluginYamlGeneration()' instead",
+        ReplaceWith("plugin.disablePluginYamlGeneration()")
+    )
     public fun disableMetaGeneration() {
-        _generateMeta = false
+        plugin.disablePluginYamlGeneration()
     }
 }